updating

+++ riboSee_pipe.sh
@@ -6,25 +6,25 @@ usage () {
   Sequences are first aligned to rRNAs and tRNAs annotated in the GTF, and then the unaligned reads are aligned to the reference genome.
   
   Mode 1:
-    -dl|--container = download the singularity container to this path and exit
+  -dl|--container = download the singularity container to this path and exit
   
   Mode 2:
-    -mt|--get_metrics = supply a dir and get metrics for all analyses in that dir, all other options will be ignored if this is non-empyt
+  -mt|--get_metrics = supply a dir and get metrics for all analyses in that dir, all other options will be ignored if this is non-empyt
   
   Mode 3: 
-    -t|--threads
-    -g|--genome_reference = path to genome reference
-    -gtf|--GTF_reference
-    -r|--reads = the fastq file
-    -o|--out_dir
-    -n|--name = output name
-    -s|--strand = stranded library (yes|no|reverse)
-    -sd|--script_directory
-    -fq|--fastQC = run fastqc?
-    -tm|--trim_fasta = path to multifasta with adapters, linkers etc to trim
-    -mm|--max_missmatch (defult = 2)
-    -mn|--min_len (defult = 24)
-    -mx|--max_len (defult = 36)
+  -t|--threads
+  -g|--genome_reference = path to genome reference
+  -gtf|--GTF_reference
+  -r|--reads = the fastq file
+  -o|--out_dir
+  -n|--name = output name
+  -s|--strand = stranded library (yes|no|reverse)
+  -sd|--script_directory
+  -fq|--fastQC = run fastqc?
+  -tm|--trim_fasta = path to multifasta with adapters, linkers etc to trim
+  -mm|--max_missmatch (defult = 2)
+  -mn|--min_len (defult = 24)
+  -mx|--max_len (defult = 36)
   "
 }
 
@@ -34,95 +34,95 @@ declare_globals () {
     # if same thing twice will take second one
     while [[ "$#" -gt 0 ]]
     do
-        case $1 in
+      case $1 in
         -t|--threads)
-        threads="$2"
-        ;;
-        -g|--genome_reference)
-        ref="$2"
-        ;;
-        -gtf|--GTF_reference)
-        gtf="$2"
-        ;;
-        -rd|--read_dir)
-        read_dir="$2"
-        ;;
-        -r|--reads) 
-        reads="$2"
-        ;;
-        -o|--out_dir) 
-        out_dir="$2"
-        ;;
-        -n|--name) 
-        nme="$2"
-        ;;
+threads="$2"
+;;
+-g|--genome_reference)
+ref="$2"
+;;
+-gtf|--GTF_reference)
+gtf="$2"
+;;
+-rd|--read_dir)
+read_dir="$2"
+;;
+-r|--reads) 
+reads="$2"
+;;
+-o|--out_dir) 
+out_dir="$2"
+;;
+-n|--name) 
+nme="$2"
+;;
         -s|--strand) #stranded library (yes|no|reverse)
-        strand="$2"
-        ;;
-        -sd|--script_directory)
-        Script_dir="$2"
-        ;;
-        -dl|--container)
-        container="$2"
-        ;;
-        -mt|--get_metrics)
-        get_metrics="$2"
-        ;;
-        -fq|--fastQC)
-        fastQC="Y"
-        ;;
-        -mm|--max_missmatch)
-        max_missmatch="$2"
-        ;;
-        -mn|--min_len)
-        min_len="$2"
-        ;;
-        -mx|--max_len)
-        max_len="$2"
-        ;;
-        -tm|--trim_fasta)
-        trim_fasta="$2"
-        ;;
-        -ca|--cut_adapt)
-        cut_adapt="$2"
-        ;;
-    esac
-        shift
-    done
+strand="$2"
+;;
+-sd|--script_directory)
+Script_dir="$2"
+;;
+-dl|--container)
+container="$2"
+;;
+-mt|--get_metrics)
+get_metrics="$2"
+;;
+-fq|--fastQC)
+fastQC="Y"
+;;
+-mm|--max_missmatch)
+max_missmatch="$2"
+;;
+-mn|--min_len)
+min_len="$2"
+;;
+-mx|--max_len)
+max_len="$2"
+;;
+-tm|--trim_fasta)
+trim_fasta="$2"
+;;
+-ca|--cut_adapt)
+cut_adapt="$2"
+;;
+esac
+shift
+done
 }
 
 
 
 if [ $# == 0 ]
 then
-    usage
-    exit 1
+  usage
+  exit 1
 fi
 
 
 ###############
 ## Functions ##
 qc_trim_SE () {
-    echo "Started trimming"
-    out_fq="${2/.f*/.trimmed.fq.gz}"
-    out_fq="$(basename $out_fq)"
-    out_fq="${5}/${out_fq}"
+  echo "Started trimming"
+  out_fq="${2/.f*/.trimmed.fq.gz}"
+  out_fq="$(basename $out_fq)"
+  out_fq="${5}/${out_fq}"
 
-    if [[ ! -z $cut_adapt ]]
-    then
-      cutadapt --cores=$1 \
-        --quality-cutoff 10,10 \
-        --minimum-length $4 \
-        -g $cut_adapt \
-        -o "${5}/${out_fq}" "$2"
+  if [[ ! -z $cut_adapt ]]
+  then
+    cutadapt --cores=$1 \
+    --quality-cutoff 10,10 \
+    --minimum-length $4 \
+    -g $cut_adapt \
+    -o "${5}/${out_fq}" "$2"
       # Regular 3’ adapter  -a ADAPTER
       # Regular 5’ adapter  -g ADAPTER
     else
       java -jar "$TRIM" SE \
-        -threads $1 \
-        "$2" \
-        "$out_fq" \
-        ILLUMINACLIP:"$3":2:30:10 LEADING:2 TRAILING:2 SLIDINGWINDOW:3:10 MINLEN:$4
+      -threads $1 \
+      "$2" \
+      "$out_fq" \
+      ILLUMINACLIP:"$3":2:30:10 LEADING:2 TRAILING:2 SLIDINGWINDOW:3:10 MINLEN:$4
     fi
 
     #FastQC post
@@ -134,7 +134,7 @@ qc_trim_SE () {
 
     reads="$out_fq"
     export reads
-}
+  }
 
 ###############
 
@@ -146,19 +146,21 @@ TRIM=/usr/bin/Trimmomatic-0.39/trimmomatic-0.39.jar
 
 if [[ ! -z $container ]]
 then
-    cd $container
-    singularity pull library://semiquant/default/riri:v0.1
-    exit 0
+  cd $container
+  singularity pull library://semiquant/default/riri:v0.1
+  exit 0
 fi
 
 if [[ ! -z $get_metrics ]]
 then
-    cd $get_metrics
-    multiqc "$get_metrics" -n $(basename $get_metrics)
-    exit 0
+  cd $get_metrics
+  multiqc "$get_metrics" -n $(basename $get_metrics)
+  exit 0
 fi
 
 # Set defults
+msk_or_rem="Nomask" #for mask set to "mask"
+
 Script_dir_tmp="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 Script_dir="${Script_dir:-$Script_dir_tmp}"
 threads="${threads:-4}"
@@ -209,31 +211,41 @@ if [ ! -f "${ref_rem}.1.ebwt" ]
 then
   # split rRNAs an tRNAs
   awk '{if ($3 == "tRNA" || $3 == "rRNA") print $0;}' "$gtf" > "${gtf/.g*/_rRNA.gtf}"
-  bedtools getfasta -fi "$ref" -bed "${gtf/.g*/_rRNA.gtf}" > "$ref_rem"
-  bowtie-build --threads $threads "$ref_rem" "$ref_rem"
+  if [[ $msk_or_rem == "mask" ]]
+  then
+    if [ ! -f "${ref/.f*/_rRNAsMasked.fasta}.1.ebwt" ]
+    then
+      bedtools maskfasta -fi "$ref" -bed test.bed -fo "${ref/.f*/_rRNAsMasked.fasta}"
+      bowtie-build --threads $threads "${ref/.f*/_rRNAsMasked.fasta}" "${ref/.f*/_rRNAsMasked.fasta}"
+    fi
+    ref="${ref/.f*/_rRNAsMasked.fasta}"
+  else
+    bedtools getfasta -fi "$ref" -bed "${gtf/.g*/_rRNA.gtf}" > "$ref_rem"
+    bowtie-build --threads $threads "$ref_rem" "$ref_rem"
+  fi
 fi
 
 
 # generate metagene `roi` file
 if [ ! -f "${ref/.f*/_rois.txt}" ]
 then
-    if [ ! -f "${gtf_ps}_5p.gtf" ]
+  if [ ! -f "${gtf_ps}_5p.gtf" ]
+  then
+    if [[ "${gtf##*.}" == "gff" ]]
     then
-        if [[ "${gtf##*.}" == "gff" ]]
-        then
-          if [ -f "${gtf/.gff/.gtf}" ]
-          then
-            gtf_ps="${gtf/.gff/.gtf}"
-          else
-            gffread "$gtf" -T -o "${gtf/.gff/_tmp.gtf}"
-            gtf_ps="${gtf/.gff/_tmp.gtf}"
-            fi
-          fi
+      if [ -f "${gtf/.gff/.gtf}" ]
+      then
+        gtf_ps="${gtf/.gff/.gtf}"
+      else
+        gffread "$gtf" -T -o "${gtf/.gff/_tmp.gtf}"
+        gtf_ps="${gtf/.gff/_tmp.gtf}"
+      fi
+    fi
         # run python script to make understandable gtf_tmp
         # this is only bcause bacterial annotations in file
         # python3 "${Script_dir}/gtf_primer.py" --gtf_in "$gtf_ps" > /dev/null 2>&1
-    fi
-    metagene generate "${ref/.f*/}" \
+  fi
+      metagene generate "${ref/.f*/}" \
       --landmark cds_start \
       --downstream 100 \
       --annotation_files "$gtf_ps"
@@ -255,16 +267,21 @@ fi
 qc_trim_SE $threads "$reads" "$trim_fasta" "$min_len" "$out_dir"
 
 # align reads
-bowtie --threads $threads --seed 1987 -x "$ref_rem" -q "$reads" -a -v $max_missmatch \
-  --un "${reads}_cleaned.fq" > /dev/null
+if [[ $msk_or_rem != "mask" ]]
+then
+  # not sure why this isnt working for all tRNAs, trying version below
+  # bowtie --threads $threads --seed 1987 -x "$ref_rem" -q "$reads" -a -v $max_missmatch \
+  # --un "${reads}_cleaned.fq" > /dev/null
+  bowtie --threads $threads --seed 1987 -x "$ref_rem" -q "$reads" -m 2 --best --strata -a -v $max_missmatch \
+    --un "${reads}_cleaned.fq" > /dev/null
+  reads="${reads}_cleaned.fq"
+fi
 
-reads="${reads}_cleaned.fq"
 bowtie --threads $threads -S --seed 1987 -x "$ref" -q "$reads" -m 5 --best --strata -a -v $max_missmatch "${out_dir}/${nme}.sam"
 samtools view -bS "${out_dir}/${nme}.sam" | samtools sort -@ $threads -O "bam" -T "working" -o "${out_dir}/${nme}.bam"
 samtools index "${out_dir}/${nme}.bam"
 rm "${out_dir}/${nme}.sam"
-samtools flagstat "${out_dir}/${nme}.bam"
-
+samtools flagstat "${out_dir}/${nme}.bam" > "${out_dir}/${nme}.flagstat"
 
 # count alignments
 echo "HtSeq started"
@@ -316,22 +333,22 @@ htseq-count --nprocesses $threads --type "CDS" --idattr "Name" --order "name" --
 #     --landmark cds_start \
 #     --annotation_files NC_000962.gtf
 phase_by_size "${ref/.f*/_rois.txt}" "${nme}_phase_by_size" \
-    --count_files "${out_dir}/${nme}.bam" \
-    --fiveprime \
-    --offset 14 \
-    --codon_buffer 5 \
-    --min_length $min_len \
-    --max_length $max_len
+  --count_files "${out_dir}/${nme}.bam" \
+  --fiveprime \
+  --offset 14 \
+  --codon_buffer 5 \
+  --min_length $min_len \
+  --max_length $max_len
 
 mkdir "${out_dir}/count_vectors"
 get_count_vectors --annotation_files "${ref/.f*/_rois.bed}" \
-    --annotation_format BED \
-    --count_files "${out_dir}/${nme}.bam" \
-    --fiveprime \
-    --offset 14 \
-    --min_length $min_len \
-    --max_length $max_len \
-    "${out_dir}/count_vectors"
+  --annotation_format BED \
+  --count_files "${out_dir}/${nme}.bam" \
+  --fiveprime \
+  --offset 14 \
+  --min_length $min_len \
+  --max_length $max_len \
+  "${out_dir}/count_vectors"
 
 # '''
 # # from plastid import BAMGenomeArray, FivePrimeMapFactory, BED_Reader, Transcript
@@ -366,19 +383,22 @@ get_count_vectors --annotation_files "${ref/.f*/_rois.bed}" \
 # '''
 
 metagene count "${ref/.f*/_rois.txt}" "${nme}_count" \
-                 --count_files "${out_dir}/${nme}.bam" \
-                 --fiveprime --offset 14 --normalize_over 30 200 \
-                 --min_counts 20 --cmap Blues
+  --count_files "${out_dir}/${nme}.bam" \
+  --fiveprime --offset 14 --normalize_over 30 200 \
+  --min_counts 20 --cmap Blues
 
 metagene chart "${nme}_counts_strt.png" \
-                "${nme}_count_metagene_profile.txt" \
-                 --landmark "start codon"
+  "${nme}_count_metagene_profile.txt" \
+  --landmark "start codon"
 
 metagene chart "${nme}_counts_peak.png" \
-                "${nme}_count_metagene_profile.txt" \
-                 --landmark "highest ribosome peak"
+  "${nme}_count_metagene_profile.txt" \
+  --landmark "highest ribosome peak"
 
-rm "$reads"
+if [[ $msk_or_rem != "mask" ]]
+then
+  rm "$reads"
+fi
 ##############        
 
 # sort GTF file
